+++ guides/CHANGELOG.md
@@ -1,1 +1,5 @@
+*   Added ```ActiveRecord::Relation#destroy_all!``` method
+    
+    *Brian John*
+
 Please check [4-1-stable](https://github.com/rails/rails/blob/4-1-stable/guides/CHANGELOG.md) for previous changes.